Prevent including source directory from include directory
Move format_impl_ptr to the include directory as it is needed by format.hpp

tests/CMakeLists.txt: If XLNT_SKIP_INTERNAL_TESTS is defined,
only add the public directory to our include path,
to ensure we do not include private files from inside our public header files.

xml_helper.hpp: moved to internal subdirectory to make it clear when we are using internal files.

format_test_suite: moved testing of internal API to format_impl_test_suite.

Fixes regression introduced by PR #115.

Fixes #118
