+++ tests/helpers/internal/xml_helper.hpp
