+++ tests/helpers/path_helper.hpp
@@ -5,8 +5,35 @@
 
 #include <xlnt/utils/exceptions.hpp>
 #include <xlnt/utils/path.hpp>
-#include <detail/utils/string_helpers.hpp>
 #include <xlnt/internal/features.hpp>
+
+#define XLNT_TEST_U8STRING_LITERAL(a) u8##a
+
+#ifdef __cpp_char8_t
+/// Casts const char8_t arrays from C++20 to const char arrays.
+inline const char * to_char_ptr(const char8_t *utf8)
+{
+  return reinterpret_cast<const char *>(utf8);
+}
+
+/// Casts char8_t arrays from C++20 to char arrays.
+inline char * to_char_ptr(char8_t *utf8)
+{
+  return reinterpret_cast<char *>(utf8);
+}
+#endif
+
+// Casts a UTF-8 string literal to a narrow string literal without changing its encoding.
+#ifdef __cpp_char8_t
+// For C++20 and newer, interpret as UTF-8 and then cast to string literal
+#define XLNT_TEST_U8_TO_CHAR_PTR(a) to_char_ptr(a)
+#else
+// For C++11, C++14 and C++17, simply interpret as UTF-8, which works with classic string literals.
+#define XLNT_TEST_U8_TO_CHAR_PTR(a) a
+#endif
+
+// The following weird cast ensures that the string is UTF-8 encoded at all costs!
+#define XLNT_TEST_ENSURE_UTF8_LITERAL(a) XLNT_TEST_U8_TO_CHAR_PTR(XLNT_TEST_U8STRING_LITERAL(a))
 
 #if XLNT_HAS_INCLUDE(<string_view>) && XLNT_HAS_FEATURE(U8_STRING_VIEW)
   #include <string_view>
@@ -29,7 +56,7 @@ class path_helper
 public:
     static xlnt::path test_data_directory()
     {
-        static const std::string data_dir = ENSURE_UTF8_LITERAL(XLNT_TEST_DATA_DIR);
+        static const std::string data_dir = XLNT_TEST_ENSURE_UTF8_LITERAL(XLNT_TEST_DATA_DIR);
         return xlnt::path(data_dir);
     }
 
@@ -47,7 +74,7 @@ public:
 
     static xlnt::path benchmark_data_directory()
     {
-        static const std::string data_dir = ENSURE_UTF8_LITERAL(XLNT_BENCHMARK_DATA_DIR);
+        static const std::string data_dir = XLNT_TEST_ENSURE_UTF8_LITERAL(XLNT_BENCHMARK_DATA_DIR);
         return xlnt::path(data_dir);
     }
 
@@ -65,7 +92,7 @@ public:
 
     static xlnt::path sample_data_directory()
     {
-        static const std::string data_dir = ENSURE_UTF8_LITERAL(XLNT_SAMPLE_DATA_DIR);
+        static const std::string data_dir = XLNT_TEST_ENSURE_UTF8_LITERAL(XLNT_SAMPLE_DATA_DIR);
         return xlnt::path(data_dir);
     }
