+++ tests/helpers/temporary_file.hpp
@@ -3,7 +3,6 @@
 #include <cstdio>
 #include <string>
 
-#include <detail/external/include_windows.hpp>
 #include <helpers/path_helper.hpp>
 
 namespace {
